requirements.txt
numpy>=1.21,<1.24
pandas>=2.1.0
scikit-learn>=1.3.0
xgboost>=2.0.1
pycaret==3.1.0
shap>=0.42.1
psycopg2-binary==2.9.9
boto3==1.29.3
python-dotenv>=1.0.0
pytest==7.4.3
plotly>=5.18.0
feature-engine>=1.6.1
imbalanced-learn>=0.11.0
tensorflow>=2.14.0
joblib>=1.3.2
fastapi>=0.104.1
uvicorn>=0.24.0
requests>=2.31.0 